secure admin endpoint action (users)

+++ api/tags.js
@@ -69,7 +69,6 @@ router.post("/filter/", async (req, res) => {
     });
   }
 });
-
 // POST new tag in DB
 router.post("/", async (req, res) => {
   if (!req.isAuth) {
@@ -78,31 +77,38 @@ router.post("/", async (req, res) => {
     });
     return;
   }
-
-  //TODO: CHeck tha user has admin rights
-
-  try {
-    const checkIfTagsExist = await client.query(
-      `SELECT * FROM tags WHERE tag_name='${req.body.tag_name}'`
-    );
-    if (checkIfTagsExist.rows.length > 0) {
-      res.status(201).json({
-        value: "failed",
-        message: `${req.body.tag_name} was already in the table 'tags'`,
-      });
-    } else {
-      await client.query(
-        `INSERT INTO public.tags (tag_name) VALUES ('${req.body.tag_name}');`
+  //Check that user has admin rights
+  const user = await client.query(`SELECT * FROM users WHERE id=${req.userId}`);
+  console.log(user.rows[0].is_admin);
+  if (!user.rows[0].is_admin) {
+    res.status(401).json({
+      error: "You do not have administrator rights.",
+    });
+    return;
+  } else {
+    try {
+      const checkIfTagsExist = await client.query(
+        `SELECT * FROM tags WHERE tag_name='${req.body.tag_name}'`
       );
-      res.status(201).json({
-        value: "success",
-        message: `${req.body.tag_name} was added to the table 'tags'`,
+      if (checkIfTagsExist.rows.length > 0) {
+        res.status(201).json({
+          value: "failed",
+          message: `${req.body.tag_name} was already in the table 'tags'`,
+        });
+      } else {
+        await client.query(
+          `INSERT INTO public.tags (tag_name) VALUES ('${req.body.tag_name}');`
+        );
+        res.status(201).json({
+          value: "success",
+          message: `${req.body.tag_name} was added to the table 'tags'`,
+        });
+      }
+    } catch (err) {
+      res.status(400).json({
+        error: `${err})`,
       });
     }
-  } catch (err) {
-    res.status(400).json({
-      error: `${err})`,
-    });
   }
 });
 
@@ -114,50 +120,52 @@ router.post("/edit/", async (req, res) => {
     });
     return;
   }
-
-  //TODO: CHeck tha user has admin rights
-
-  try {
-    const oldtag = req.body.oldtag;
-    const newtag = req.body.newtag;
-
-    // Rename in picture all from oldtag to newtag
-    const selectPictureWithTagsQuery = `SELECT id, tags FROM pictures WHERE tags LIKE '%"${oldtag}"%'`;
-    const pictureWithTag = await client.query(selectPictureWithTagsQuery);
-    if (pictureWithTag.rows.length < 1) {
+  //Check that user has admin rights
+  const user = await client.query(`SELECT * FROM users WHERE id=${req.userId}`);
+  console.log(user.rows[0].is_admin);
+  if (!user.rows[0].is_admin) {
+    res.status(401).json({
+      error: "You do not have administrator rights.",
+    });
+    return;
+  } else {
+    try {
+      const oldtag = req.body.oldtag;
+      const newtag = req.body.newtag;
+      // Rename in picture all from oldtag to newtag
+      const selectPictureWithTagsQuery = `SELECT id, tags FROM pictures WHERE tags LIKE '%"${oldtag}"%'`;
+      const pictureWithTag = await client.query(selectPictureWithTagsQuery);
+      if (pictureWithTag.rows.length < 1) {
+        res.status(400).json({
+          error: `No picture found with the tag '${oldtag}'`,
+        });
+        return;
+      } else {
+        pictureWithTag.rows.forEach(async (row) => {
+          const updatedTags = row.tags.replace(oldtag, newtag);
+          const updateTagQuery = `UPDATE pictures SET tags='${updatedTags}' WHERE id=${row.id};`;
+          await client.query(updateTagQuery);
+        });
+      }
+      // Check if new Tag exist in the tag_table already
+      const createTagQuery = `SELECT * FROM tags WHERE tag_name='${newtag}'`;
+      const doesTagExist = await client.query(createTagQuery);
+      if (doesTagExist.rows.length < 1) {
+        // if no, Create new Tag in tag_table
+        const createTagQuery = `INSERT INTO tags (tag_name) VALUES ('${newtag}');`;
+        await client.query(createTagQuery);
+      }
+      // Delete the old tag form the tag_table
+      const deleteTagQuery = `DELETE FROM tags WHERE tag_name='${oldtag}';`;
+      await client.query(deleteTagQuery);
+      res.status(201).json({
+        message: `Tag '${oldtag}' was updated to '${newtag}'`,
+      });
+    } catch (err) {
       res.status(400).json({
-        error: `No picture found with the tag '${oldtag}'`,
-      });
-      return;
-    } else {
-      pictureWithTag.rows.forEach(async (row) => {
-        const updatedTags = row.tags.replace(oldtag, newtag);
-        const updateTagQuery = `UPDATE pictures SET tags='${updatedTags}' WHERE id=${row.id};`;
-        await client.query(updateTagQuery);
+        error: `${err})`,
       });
     }
-
-    // Check if new Tag exist in the tag_table already
-    const createTagQuery = `SELECT * FROM tags WHERE tag_name='${newtag}'`;
-    const doesTagExist = await client.query(createTagQuery);
-
-    if (doesTagExist.rows.length < 1) {
-      // if no, Create new Tag in tag_table
-      const createTagQuery = `INSERT INTO tags (tag_name) VALUES ('${newtag}');`;
-      await client.query(createTagQuery);
-    }
-
-    // Delete the old tag form the tag_table
-    const deleteTagQuery = `DELETE FROM tags WHERE tag_name='${oldtag}';`;
-    await client.query(deleteTagQuery);
-
-    res.status(201).json({
-      message: `Tag '${oldtag}' was updated to '${newtag}'`,
-    });
-  } catch (err) {
-    res.status(400).json({
-      error: `${err})`,
-    });
   }
 });
 
@@ -169,40 +177,46 @@ router.delete("/", async (req, res) => {
     });
     return;
   }
+  //Check that user has admin rights
+  const user = await client.query(`SELECT * FROM users WHERE id=${req.userId}`);
+  console.log(user.rows[0].is_admin);
+  if (!user.rows[0].is_admin) {
+    res.status(401).json({
+      error: "You do not have administrator rights.",
+    });
+    return;
+  } else {
+    try {
+      const tagToDelete = req.body.tagToDelete;
+      // Delete tagToDelete from tag
+      const deleteTagQuery = `DELETE FROM tags WHERE tag_name='${tagToDelete}';`;
+      console.log(deleteTagQuery);
+      await client.query(deleteTagQuery);
 
-  //TODO: CHeck tha user has admin rights
-
-  try {
-    const tagToDelete = req.body.tagToDelete;
-    // Delete tagToDelete from tag
-    const deleteTagQuery = `DELETE FROM tags WHERE tag_name='${tagToDelete}';`;
-    console.log(deleteTagQuery);
-    await client.query(deleteTagQuery);
-
-    // Delete Tag in picture
-    const selectPictureWithTagsQuery = `SELECT id, tags FROM pictures WHERE tags LIKE '%"${tagToDelete}"%'`;
-    const pictureWithTag = await client.query(selectPictureWithTagsQuery);
-    if (pictureWithTag.rows.length < 1) {
-      res.status(400).json({
-        error: `No picture found with the tag '${tagToDelete}'`,
+      // Delete Tag in picture
+      const selectPictureWithTagsQuery = `SELECT id, tags FROM pictures WHERE tags LIKE '%"${tagToDelete}"%'`;
+      const pictureWithTag = await client.query(selectPictureWithTagsQuery);
+      if (pictureWithTag.rows.length < 1) {
+        res.status(400).json({
+          error: `No picture found with the tag '${tagToDelete}'`,
+        });
+        return;
+      } else {
+        pictureWithTag.rows.forEach(async (row) => {
+          const updatedTags = row.tags.replace(`"${tagToDelete}"`, null);
+          const updateTagQuery = `UPDATE pictures SET tags='${updatedTags}' WHERE id=${row.id};`;
+          await client.query(updateTagQuery);
+        });
+      }
+      res.status(201).json({
+        value: "success",
+        message: `Tag '${tagToDelete}' was deleted.`,
       });
-      return;
-    } else {
-      pictureWithTag.rows.forEach(async (row) => {
-        const updatedTags = row.tags.replace(`"${tagToDelete}"`, null);
-        const updateTagQuery = `UPDATE pictures SET tags='${updatedTags}' WHERE id=${row.id};`;
-        await client.query(updateTagQuery);
+    } catch (err) {
+      res.status(400).json({
+        error: `${err})`,
       });
     }
-
-    res.status(201).json({
-      value: "success",
-      message: `Tag '${tagToDelete}' was deleted.`,
-    });
-  } catch (err) {
-    res.status(400).json({
-      error: `${err})`,
-    });
   }
 });
